ajout du buttonguard pour sécuriser les actions tels que le vote,follow,messaging : structuration designe du profil form

+++ src/app/category/common/category-common.component.ts
@@ -1,5 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { CategoryService } from '../service/category.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'category-common',
@@ -11,11 +12,14 @@ export class CategoryCommonComponent {
   @Input()
   category : any;
 
-  constructor(private categoryService : CategoryService) {}
+  constructor(private categoryService : CategoryService,
+              private router: Router) {}
 
   postCategoryFollow(category_id: any){
-    this.category.is_follow = true;
-    this.categoryService.postCategoryFollow(category_id).subscribe(
+    
+    if(this.buttonGuard() == true){
+      this.category.is_follow = true;
+      this.categoryService.postCategoryFollow(category_id).subscribe(
         data => {
           //illusion optiques
         },
@@ -24,12 +28,17 @@ export class CategoryCommonComponent {
           this.category.is_follow = false;
         },
         () => console.log("finish")
-    );
+      );
+    }else{
+       this.buttonGuardRedirect();
+    }
+
   }
   
   removeCategoryFollow(category_id: any){
-    this.category.is_follow = false;
-    this.categoryService.deleteCategoryFollow(category_id).subscribe(
+    if(this.buttonGuard() == true){
+      this.category.is_follow = false;
+      this.categoryService.deleteCategoryFollow(category_id).subscribe(
         data => {
           //illusion optique
         },
@@ -38,7 +47,26 @@ export class CategoryCommonComponent {
           this.category.is_follow = true;
         },
         () => console.log("finish")
-    );
+      );
+    }else{
+       this.buttonGuardRedirect();
+    }
+    
   }
+
+  buttonGuard(){
+    if(localStorage.getItem("access_token")){
+      return true;
+    }else{
+      return false;
+    }
+  }
+ 
+ buttonGuardRedirect(){
+   if(window.confirm('Voulez-vous vous connecter pour effectuer cette action ?'))
+    {
+      this.router.navigateByUrl('/login');
+    }
+ }
 
 }
